Test 6 : Backend code refactoring (Success)
