+++ frontend-react-native/components/StatsData.js
@@ -7,7 +7,7 @@ import axios from 'axios';
 export default DataContainer = () => {
     let [stats, setStats] = useState([]);
     let compteur = 0;
-    const url = 'http://172.20.10.6:8000/';
+    const url = 'http://192.168.1.56:8000/';
 
     useEffect(() => {
         getStats();
